insurance-claims-backend/insurance-claims-backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
pymongo
langgraph
networkx
numpy
pandas
scikit-learn
xgboost
pytest
python-dotenv








# Core
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.3.0

# Database
motor==3.5.0
pymongo==4.8.0

# Web Framework (for integration later)
fastapi==0.115.0
uvicorn[standard]==0.30.0

# LangGraph (Open Source)
langgraph==0.2.0
langchain-core==0.3.0

# ML - All Open Source
scikit-learn==1.5.0
xgboost==2.1.0
numpy==1.26.0
pandas==2.2.0
openml==0.14.2
joblib==1.4.0

# Text Processing (Free)
fuzzywuzzy==0.18.0
python-levenshtein==0.25.0  # Optional, speeds up fuzzywuzzy
rapidfuzz==3.6.1

# Embeddings (Free, runs locally)
sentence-transformers==3.0.0

# Utilities
python-dateutil==2.9.0



# requirements.txt - Complete Dependencies for Node 3 & Node 4

# ============================================================
# CORE DEPENDENCIES (Required for both Node 3 & Node 4)
# ============================================================

# FastAPI & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
motor==3.5.0
pymongo==4.8.0

# Environment & Configuration
python-dotenv==1.0.1

# Data Validation
pydantic==2.7.0
pydantic-settings==2.3.0

# Date/Time Handling
python-dateutil==2.9.0

# ============================================================
# NODE 3 DEPENDENCIES (Policy Coverage Agent)
# ============================================================

# No additional dependencies beyond core
# Node 3 uses only Python standard library + core

# ============================================================
# NODE 4 DEPENDENCIES (Fraud Detection Agent)
# ============================================================

# === 1. ANOMALY DETECTION ===
# Isolation Forest, ML models
scikit-learn==1.5.0
xgboost==2.1.0
numpy==1.26.0
pandas==2.2.0
joblib==1.4.2

# === 2. DOCUMENT COMPARISON ===
# RapidFuzz is 10x faster than FuzzyWuzzy
rapidfuzz==3.9.0
# Optional: python-levenshtein for even faster Levenshtein distance
python-levenshtein==0.25.0

# === 3. NETWORK ANALYSIS ===
# Graph algorithms for fraud ring detection
networkx==3.3
# Optional: For larger graphs, use iGraph
# python-igraph==0.11.0

# === 4. EXPLAINABLE AI ===
# SHAP for model explanations
shap==0.45.0
# Optional: LIME as alternative
# lime==0.2.0.1

# === 5. TEXT PROCESSING (Optional) ===
# For advanced text analysis in documents
# nltk==3.8.1
# spacy==3.7.4

# ============================================================
# TESTING DEPENDENCIES
# ============================================================

pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# ============================================================
# DEVELOPMENT TOOLS (Optional)
# ============================================================

# Code formatting
black==24.4.0
# Linting
ruff==0.4.0
# Type checking
mypy==1.10.0